Re-organized unit test folder and files
- one file with 'main' plus necessary global SOILWAT2 variables -->
'test/sw_maintest.cc'
- renamed file 'test/sw_test.cc' to 'test/test_rands.cc' which contains
all unit tests for functionality from the rands.c/rands.h files

-> additional unit tests can be written and added their own file like
'test_rands.cc'
