+++ test/sw_maintest.cc
@@ -1,0 +1,30 @@
+#include "gtest/gtest.h"
+#include <assert.h>
+#include <ctype.h>
+#include <dirent.h>
+#include <errno.h>
+#include <float.h>
+#include <math.h>
+#include <memory.h>
+#include <stdarg.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <sys/stat.h>
+#include <sys/types.h>
+#include <time.h>
+#include <unistd.h>
+#include "../generic.h"
+
+
+
+char errstr[MAX_ERROR]; /* used to compose an error msg    */
+FILE *logfp; /* file handle for logging messages */
+int logged; /* boolean: true = we logged a msg */
+
+
+int main(int argc, char **argv) {
+  ::testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
+}
+
