+++ test/test_rands.cc
@@ -23,13 +23,8 @@
 #include "../rands.h"
 #include "../rands.c"
 
-char errstr[MAX_ERROR]; /* used to compose an error msg    */
-FILE *logfp; /* file handle for logging messages */
-int logged; /* boolean: true = we logged a msg */
-
-
 namespace {
-  // This test belongs to the beta random number generator
+  // This tests the beta random number generator
   TEST(BetaGeneratorTest, ZeroToOneOutput) {
     EXPECT_LT(RandBeta(0.5, 2), 1);
     EXPECT_LT(RandBeta(1, 3), 1);
@@ -43,19 +38,4 @@ namespace {
     EXPECT_DEATH(RandBeta(-1, -3), "AA <= 0.0");
   }
 
-
-TEST(IsTestTest, Failure) {
-  // This test belongs to the IsTestTest test case.
-
-  EXPECT_FALSE(3 == 4);
-}
-
 } // namespace
-
-
-int main(int argc, char **argv) {
-  ::testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-
-
-}
